backend/requirements.txt
alembic==1.14.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bcrypt==4.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Mako==1.3.9
MarkupSafe==3.0.2
nltk==3.9.1
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.3
python-engineio==4.11.2
python-socketio==5.12.1
regex==2024.9.11
requests==2.32.3
simple-websocket==1.1.0
SQLAlchemy==2.0.38
sqlparse==0.5.2
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
